Move criteria to right of status chooser when there's room (fixes #862) (#967)
When there's enough width, move the criteria text to the right of
the chooser.  The justification text continues to be below the
chooser, to maximize the amount of space available for long answers.

We originally moved the criteria text to a new line because the Russian text
was too long in the original format.  That solved one problem,
but created a new one.

As @KitsuneRal noted, "I have to say though that criteria take a lot
of vertical screen estate now that the radio group is displayed in
a separate div from the text and both are float: left. And since the
current representation doesn't map to the conventional "question-response"
arrangement, I already caught myself once mistakenly associating the radio
group with a criteria above, not below - though the (barely visible on
poorer monitors) separator line tries to mitigate that."

Now that the criteria text and the responses are on the same line
(on larger monitors), this confusion should disappear.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>