Tweak Rakefile: regenerate doc/criteria.md if criteria text changes
When config/locales/en.yml changes, be sure to regenerate doc/criteria.md.
At one time all criteria text was in criteria/criteria.yml,
so we didn't need anything else.  However, as part of i18n we moved
a lot of text to config/locales/en.yml, but forgot to add this as
a dependency.  This commit fixes that.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>