Add 'tuf' project to front page logos
Add The Update Framework (tuf) to the set of front-page projects
with logos.  They just earned a badge (congrats!), and are a different
kind of security project covering an important issue (software updates).
We want to promote this kind of thing, so pointing to them makes sense.
For more about tuf, see: <https://www.theupdateframework.com>

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>