Clarify: projects must use https: for sites_https
We're long required that projects must use https: for their project
and version control repos to meet criterion sites_https.
This commit makes that more obvious, by explaining this in the DETAILS.
We also strongly encourage projects to use https (not http) when
they create their project entry, which should reduce the number
of times this problem occurs.

This change was inspired by issue #991.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>