Require rationale to delete project and report it
Currently people sometimes delete project badge entries and
we have no idea why.  Also, although we required confirmation
before deletion, it was still very easy for people to
accidentally delete project entries.

This commit changes the "delete" UI link to instead present
a special form (delete_form),
where we "require" a deletion rationale.
That deletion rationale is included in the email to admins
that occurs when a project is deleted.  This makes it
much less likely that a deletion will happen by accident,
and the rationale may help us better understand why some
people delete project entries.

We only use HTML to "require" the deletion rationale, and don't
bother to check that on the server, since failing to provide
this information isn't a security issue.  We just would like to know.

This uses a separate "delete_form" name and method,
not "delete", since Rails handles "delete" specially.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>