Enable client-side queries using CORS, fixes #995
In issue #995 Tony Hansen reported that client-side
JavaScript that originates anywhere else
cannot query for information (e.g., using getJSON),
even though this works fine from curl, etc.

The problem is that the Same Origin Policy, as implemented
in all client-side JavaScript implementations, by default disallows reading
a remote resource with a different origin.

This commit adds Cross-Origin Request Sharing (CORS)
so JavaScript clients can GET data from the BadgeApplication
(in particular, can get JSON data).
In particular, we add an "Access-Control-Allow-Origin" header
to the set of HTTP headers if there is an origin.

Note that we *ONLY* relax the same origin policy on "GET",
and we do NOT share credentials.  The not-sharing-credentials
is the key part; since the different origin doesn't get the keys
(and isn't allowed to use them), more-privileged actions are simply
not permitted via CORS unless the original requester already has those
privileges.  For more on CORS, see:
https://en.wikipedia.org/wiki/Cross-origin_resource_sharing

The "secure_headers" gem from Twitter manages other HTTP header
security configurations, but it does not directly support CORS
(see its issue #331).

Instead, we add the Rack gem cyu/rack-cors, as this seems to be the
normal way CORS handling is done in Rails.

Oversharing credentials via CORS is a common problem.
For more about this, see:
"Exploiting CORS Misconfigurations for Bitcoins and Bounties" by James Kettle
http://blog.portswigger.net/2016/10/exploiting-cors-misconfigurations-for.html
https://www.youtube.com/watch?
  v=wgkj4ZgxI4c&list=PLpr-xdpM8wG8RHOguwOZhUHkKiDeWpvFp&index=31

Note its issue 126, "Rack CORS is fundamentally insecure",
which warns that it is notoriously easy to misconfigure things
to share credentials.

To counteract oversharing, the new tests include a check
to ensure that when CORS is enabled, we do *not* enable credential sharing.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>