Test to ensure caching varies depending on CORS origin
It could be a security disaster if caches always served the same data
to different CORS origins, because private information (e.g., to system
admins) might leak in a cache.  The "rack-cors" gem already handles
this by inserting "Vary: Origin", as is standard recommended behavior
when using CORS.  However, it's important for security, so let's
verify that caching actually varies depending on the Origin.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>