Add support for /projects/:id/badge.json and include badge_level in JSON
We already supported /projects/:id/badge and by default that is SVG.
Add support for /projects/:id/badge.json so that external queries
can easily determine the badge level of a project.  The reply
includes "badge_level" which is one of the short badge level names
("in_progress", "passing", "silver", or "gold").

For consistency, also provide badge_level when replying with a full
JSON response for a project (e.g., /projects/:id.json).

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>