Add CORS library middleware stack location checks
Add tests to check the CORS library configuration.

Because of the way it works, Rack::Cors *must* be first in the Rack
middleware stack, as documented here: https://github.com/cyu/rack-cors
These tests verify this precondition, because it'd be easy to
accidentally cause this assumption to fail as code is changed and
gems are added or updated.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>