Change justification comment mark from '#' to '//' fixes #1009
Up to now, if a justification textbox begins with "#", optionally
preceded by spaces, it is considered a "justification comment"
(and is not considered a justification for purposes of an Unmet SHOULD
or where a justification is otherwise required). That made sense because
in many programming languages, "#" introduces a comment that ends at
the end of the line.

In retrospect, that was a mistake, because "#" also has a meaning in
markdown - the beginning of a header. This means that justification
comments look really wonky when viewed (though they look fine while
they're being edited).

We need to change things somehow. The simplest approach, taken here,
is to switch to leading "//" (possibly preceded by spaces) as the
marker for a justification comment. This is a comment marker in many
programming languages, including JavaScript, Java, C#, C++, and C.
Since many OSS project leaders/participants will know at least one of
these programming languages, it should be relatively clear
what's indicated. The sequence "//" is extremely unlikely to be
used accidentally, has no markdown meaning, and is easy to type, too.

Unlike the previous implementation, the text must start with
exactly "//" (even leading spaces disable it).
This is simpler and faster.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>