Clarify in README that dependency licenses can be different
Clearly state in the README that the components we depend on
do not necessarily have the MIT license.
We release *our* software under the MIT license,
but that does not affect the licenses of any
software components we depend on.  That should be obvious,
but let's make that expressly clear so there's no confusion.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>