Fix some style issues for rubocop 0.52.1
We currently use rubocop version 0.51.0.  Rubocop 0.52.1
produces many more reports, so we can't just simply switch to it.
We could disable new checkers, but we generally try to accept the
changes first, unless it's obviously a bad idea.

This commit fixes many of the "enable after disable" messages, and
gets us a little closer to using rubocop 0.52.1.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>