Capture more daily statistics
Capture more daily statistics.  In particular, up to this point we
have not captured *any* statistics involving users, and we haven't
captured any additional rights data either.  This commit adds more
statistics to make it easier to gain insights in how things are going.

This commit also tweaks how the project_stats JSON file is generated,
which simplifies and future-proofs the code.  Now the JSON created for
project stats simply includes "all attibutes available in project_stats".

Technically this means that the "project_stats" model, table, and URL
is misnamed, as these are really "daily_stats" now.
But that name change would be a big pain for little gain.
One problem is that many people already refer to "/project_stats".
If we want to do that, that can be done as a separate later commit.
Let's instead focus on capturing the data first (while we can).

This commit does not add generating a *graph* of this data.
We can do that later, once we have data to show.  We used the same process
for previous additions to our stats (focus on capturing data first).

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>