Enable users to delete their account and show email (GDPR) (#1025)
Change how we handle users so that users can delete their own accounts
and show their own email addresses.  The purpose of these changes is
(in part) to comply with the EU General Data Protection Regulation (GDPR)
as we understand it.

This commit enables users to delete their own user accounts to support the
EU 'right of erasure'.  The application requires that every project
entry have an 'owner', so we only allow user account deletion if
a user owns no project entries.  Since users can always delete project
entries if they own them, users can always delete their accounts without
any action from an administrator.

This commit also enables users to show their own email address (in both
the HTML and JSON format).  This supports the GDPR "right of access"
where users must be able to see the personal data we record about them.
With this change, we believe users can view all the personal data we
store about them.  (We intentionally don't store much, but what we
have we show).  This does mean that if users display a screenshot
of their page to others, it will reveal their email address to others
(a problem our previous version didn't have), but we need to comply
with GDPR and this seemed like the most obvious solution.

The JSON format for user data supports the GDPR "data portability"
requirements, since JSON is a widely-used standard format.

JSON stands for "JavaScript Object Notation (JSON)" and is defined in
"The JavaScript Object Notation (JSON) Data Interchange Format"
by T. Bray, Ed., Internet Engineering Task Force (IETF)
Request for Comments (RFC) 8259, December 2017,
<https://tools.ietf.org/html/rfc8259>
It is also defined by
Ecma International's "The JSON Data Interchange Format",
Standard ECMA-404,
<http://www.ecma-international.org/publications/standards/Ecma-404.htm>.

A number of additional tests have been added and honed.
When we limited user account deletion and email display to trusted admins,
then there were fewer cases to check.  Now that normal users can display
their own data, but normal users are not supposed to be able to see some
data about other users, we have more test cases to check on.

As hardening measures, we intentionally set the user account
pages to "noindex", and caching is also disabled on the user page.
Hyperlinks to the user from a project or project index
are marked as "nofollow".
This will reduce the set of crawled user data
available via other sources, and also reduce the impact via caches,
if there is a serious defect elsewhere in the code.
It's best to not have a problem, but it's also wise to
take reasonable prudent measures to reduce the impact of a problem,
just in case.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>