Update app/views/project_stats/show.json.jbuilder
Update file app/views/project_stats/show.json.jbuilder to match
file app/views/project_stats/index.json.jbuilder.  There's some
intentional non-DRYness here for speed; we have a loop-in-loop
with a fair amount of data, so it makes sense to directly express
index.  Also, at this point changes to this code are unlikely
(it automatically adjusts to whatever data is available).
That said, cross-reference comments have been added so that
if there *is* a change in the future to either file, there
will be ample warning.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>