Put User#destroy in a transaction (#1032)
Fix a highly-improbable race condition when deleting a user.

This commit puts the main logic of User#destroy within a transaction.
Otherwise, it is theoretically possible that between the test
"does the user own 0 projects" and "delete the user" someone could
add a project owned by the user being destroyed.

This problem is improbable, since only admins and that user can delete
the user or create a project owned by a user, and the race would have
to be timed perfectly.  There's nothing to be gained by the race, either;
it doesn't give anyone special privileges or break security.

However, it's trivial to put this inside a transaction, and I *really*
don't like having known race conditions (because they're irreproducible
they're notoriously hard to debug).

I put this in a nested transaction. As long as the tables are in the
same database there's no need for that, but if the tables were later
separated into different databases a single transaction call would
create a subtle hard-to-find bug. So let's just put both relevant
tables in the transaction, as a defensive coding step.
Nobody cares about the speed of user account deletion, so nesting
a transaction is not a big deal.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>