robots.txt: Add comments, disallow project *.json files
Insert more comments that end up in robots.txt itself, so
that anyone viewing the robots.txt file (a public file)
will understand why it has the values it does.
In particular, we tell crawlers to skip public user data,
so if users delete their accounts, that data will more rapidly
disappear elsewhere too.

In addition, tell crawlers to skip project .json files.
They're unlikely to be of value to a crawler, and we reduce
the work our application has to do by telling crawlers to not bother.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>