Add server-side check of project deletion_rationale, fixes #1022 (#1036)
Perform a server-side check of the deletion_rationale for projects,
making the check non-bypassable.  This isn't necessary for security,
but it ensures we get *some* text from the user. Hopefully the text is
useful, and in any case this provides a stronger defense against
accidental deletion.  This commit also adds some tests to ensure it works.

The proposed rule is "at least 20 characters and at least 15
non-whitespace characters".  That will at least prevent someone
accidentally providing a bunch of returns.  If people really insist
on giving us garbage (e.g., a bunch of "a"s) we don't
want them involved anyway.

The check isn't require for admins, who presumably know what they're
doing (and can fix it if they make an error).  That way, if an admin
has to delete a bunch of garbage projects, it's not *too* hard to do it.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>