Modify robots.txt to disallow /users (fixes #1027) (#1028)
* Modify robots.txt to disallow /users (fixes #1027)

This modifies the BadgeApp's "robots.txt" to disallow crawling
(and thus indexing) of anything in /users, including /LOCALE/users.

The GDPR is obviously very sensitive about user data, and we display
some user data. If another site crawls/scrapes/spiders the data (say for a
web search), then I think it's primarily the problem of the other site to
ensure that the data disappears in a reasonable period of time.

That said, we can make it much less likely that user data will show up
in spidered data at all, by expressly telling spiders to not look at
those pages. Then, if a user account is deleted, that user data will
disappear much more quickly. In addition, if we make a mistake and
accidentally provide private data (primarily the email address), it won't
be indexed in other places in most cases - reducing the damage caused
if we make a mistake. Disallowing in robots.txt removes the user pages
from search results, but I think it's unlikely people would normally
want to see our user pages as search results, and in a separate pull
request I propose setting those pages to "noindex" anyway (which also
removes them from search results). Using robots.txt will also counter
the web crawling of user pages that, while not a big deal, causes us
to do some work that is unlikely to help anyone.

See the robotstxt.org specification <http://www.robotstxt.org/orig.html>
and the Google robots.txt specification at
<https://developers.google.com/search/reference/robots_txt>.
The latter describes how Google does things and adds more capabilities,
in particular, Google has "Allow", and if there's a conflict, the longest
(most specific rule) wins.  So for example, "/users" is longer than "/"
so something matching "/users" would be disallowed.
We shouldn't specify "/users/" since both /users and /users/ map to
"index" (and we need the rules to cover both, so just using "/users"
is simpler). There are no resources starting with "/users" that aren't
user-related, so it's simplest to just disallow "/users" (and so on).

This change takes many steps to ensure that adding a locale
won't suddenly allow crawling a page.
This change modifies our instructions to ourselves on how to create
locales, AND creates a cross-reference comment at the very line you
have to change to add a locale.   In addition, the automated tests
specifically check to ensure that we disallow every locale, and those
are run before any deployment.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>

* robots.txt: Add comments, disallow project *.json files

Insert more comments that end up in robots.txt itself, so
that anyone viewing the robots.txt file (a public file)
will understand why it has the values it does.
In particular, we tell crawlers to skip public user data,
so if users delete their accounts, that data will more rapidly
disappear elsewhere too.

In addition, tell crawlers to skip project .json files.
They're unlikely to be of value to a crawler, and we reduce
the work our application has to do by telling crawlers to not bother.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>