On project create, rm trailing / in URLs
Remove trailing "/" in URLs when trying to create a project entry.

Most URLs that people submit do not have trailing "/", but
a non-trivial number do have a trailing "/", and this makes it
a little harder to notice duplicates.  Remove the trailing "/"
on initial create.

This doesn't prevent later changes, especially of the homepage_url,
but it reduces the number of initial problems.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>