Display "Logged in" not "Signed in"
Try to display "log in" and "log out" instead of
"sign in" and "sign out", so that they are more
visually distinct from "sign up".

In English the terms "sign in" and sign up" have different
yet related meanings.  "Sign up" involves registration, while
"sign in" has the same meaning as "log in".
These are standard terms, indeed, font awesome "sign-in"
and "sign-out" for login and logout.

However, the visual similarity of "sign in" and "sign up"
can be confusing.  This commit changes even more displays
so that we use "log in" and not "sign in", making
"sign up" more distinct.  We have generally done this
elsewhere, but this commit fixes a few stragglers.

It's likely that translators won't have anything to do
with this change; this is probably specific to English.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>