Preload font awesome fonts
Speed up first-time display by hinting to the web brower
to download the font awesome fonts.  The browser will eventually
figure that out, but only after it parses the CSS and more of the HTML.
Include the hint in the HTML head, so that the browser can
initiate the request sooner.

This doesn't matter after the first-time display, because the
fonts will almost certainly be cached, but this is an easy
optimization for first-time display.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>