Tweak font-awesome preload
The article "Preload: What Is It Good For?"
https://www.smashingmagazine.com/2016/02/preload-what-is-it-good-for/
Notes some issues when preloading fonts.

First, we must "add a crossorigin attribute when fetching fonts, as they
are fetched using anonymous mode CORS. Yes, even if your fonts are on
the same origin as the page. Sorry." See:
https://drafts.csswg.org/css-fonts/#font-fetching-requirements

Second, it's wise to indicate the type, so that the preload only
occurs if the browser supports the file type being loaded.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>