Preload logo
Optimize for first-time view of pages on this site.
The logo shown on the top left-hand-side is *always* shown
in HTML, however, it takes a moment for the browser to realize this
(it's part of the "logo" CSS class, and thus the browser
has to download and parse the CSS to figure this out).
Tell the browser to preload it, so that on first view the
viewer will grab it much sooner.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>