Use more font awesome and tweak signup text (#1049)
* Use more font awesome and tweak signup text

This commit adds icons to all major navigation-level actions,
and switches the ones that existed to all use font awesome.

This improves the look, and is a step towards eliminating
glyphicons and subsetting font awesome (to speed first-time display).

It also tweaks the login and signup text to make it clearer that
logging in via GitHub does an automatic sign-up.

One challenge is that in English "sign in" and "sign up"
are different operations, but they sound similar and are related,
so it's easy for people to be confused.
In addition, because we want users to have *convenient* experience,
users who sign in using GitHub also have an automatic sign up
(if necessary).  We tried to be clear before - hopefully these
changes will make it even clearer.

Since icon tags show up in many places, create a helper to generate
the tag and use it.  Putting the icon tag creation in one place
should make it easier to change icon generation later.
This new routine gets called a lot.  Optimize its use of
SafeBuffer so it doesn't create lots of transient temporaries.

Here we try to keep the UI terminology consistent, by saying
"custom" everywhere.

We also tweak some other related text.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>