Document our turbolinks experience
Document turbolinks.
I might be willing to re-add turbolinks in the future for
performance, but clearly just "sprinkling it in" doesn't work.
You really need to modify your JavaScript software to work with it.
It's not clear we need it, but let's record that information
so that the information is not lost.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>