Minor code cleanup of /project_stats chart code
Clean up code to simplify slightly.
This also adds some tests, as the process of cleaning up the
code revealed that the tests needed to be improved.
This commit adds a test for the "as admin" case of seeing
project stats, and adds simple checks of the contents of the response.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>