Fix I18n for validation messages. Fixes #1062
This commit fixes a bug so some validation messages are
properly internationalized.  The previous version *looked*
right, because it called I18n.t, but this call was executed
at *startup* time (where the locale was "en"), instead of
at the time the validation is made.

To fix this, switch to using a *symbol* for the message,
and move the error message to where Rails expects such
messages to be.  For more information, see:
https://stackoverflow.com/questions/17935597/
ruby-on-rails-i18n-want-to-translate-custom-messages-in-models

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>