Tweak robots.txt further - forbid most queries
In robots.txt, tell crawlers to *not* crawl most queries
(e.g., pq= or lteq=).  That gets rid of some of the problems
in allowing crawling of JSON files, so allow crawlers to
crawl the JSON files.

This should eliminate the current problem that crawlers tend to
stay very busy on our site with useless crawls - we don't
want crawlers to waste their time, but instead want to help
them crawl the *important* parts of the site.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>