Split too-long string, fixes #1066
KitsuneRal reported the following problem:

"sessions.login_with_github_html has become too long with a recent
addition about automatic sign-up. Phone screens have no chance to fit it,
and even on my laptop, unless I make a browser window wider than circa
1280px, the text doesn't fit in full. This addition looks like a comment
or a note anyway, so probably put it under the login 'button'?  (And yes,
with Russian I struggle even more because the translation is longer.)"

This commit fixes the problem as suggested, by moving the
"rest of the text" down below the button.

The text was made longer in an effort to help make it clearer
to users that login via GitHub can, in some cases, also cause a sign in.
This should be obvious anyway, since (1) it's common practice and
(2) GitHub will make the user specifically permit it.  In this
commit we still provide an *extra* (third) mechanism for the user to
learn this, without making the button unreadable on small screens.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>