Update to ruby 2.5.0; required vcr_cassettes reset
This updates to Ruby 2.5.0, the latest officially-released
version of Ruby.  It is supposed to be slightly faster.

This commit required updating all the vcr_cassettes, and I had
to fix a few things to make that work (it's been a while since
we updated vcr cassettes).  For example, the tests now skip loading
a VCR file if it isn't there (this is necessary when we delete
VCR files).  This commit modifies the test documentation
in doc/testing.md to explain some key points, so that this
process will be easier to repeat in the future.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>