Update to ruby 2.5.0; required vcr_cassettes reset (#1068)
* Update to ruby 2.5.0; required vcr_cassettes reset

This updates to Ruby 2.5.0, the latest officially-released
version of Ruby.  It is supposed to be slightly faster.

This commit required updating all the vcr_cassettes, and I had
to fix a few things to make that work (it's been a while since
we updated vcr cassettes).  For example, the tests now skip loading
a VCR file if it isn't there (this is necessary when we delete
VCR files).  This commit modifies the test documentation
in doc/testing.md to explain some key points, so that this
process will be easier to repeat in the future.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>

* Update gem rubocop (0.51.0->0.52.1)

This updates the gem "rubocop" to 0.52.1.  This has many new cops,
and we don't currently comply with them all, so this commit
also disables many new cops.  See issue #1069 which proposes to
do the work to comply with the new cops or reject them.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>

* Tweak circleci configuration

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>

* Add "gem update --system" per CircleCI suggestion

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>