Fix some locale test cases
This fixes some of the test cases for the change in behavior.
Note that now we generally need to change
"root_url" to "controller.root_url", and
"root_path" to "controller.root_path", because we override
root_url and root_path in the application controller (and thus
must use the overridden one, if we use one, in the test cases).

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>