Auto-locale: Reroute no-locale URLs below "/"
If no locale is provided in the URL, and it's routed by the
Rails router, then reroute to a locale...  *even* if it's not the top ("/").
This doesn't affect some paths, such as "/assets".

This is a small change with a big impact on our tests,
and causes a lot of our tests to fail.  In almost all cases
locales are now required OR the test must be prepared to redirect.

It's not clear what status code to provide on a locale-based redirect.
However, we must avoid 301 (Moved Permanently), because it is certainly
not a permanent move.  For the moment we'll use 300 (Multiple Choices),
because that code indicates there's a redirect based on agent choices
(which is certainly true).  We have to set the status to *something*,
because the Rails default is 301 and that is absolutely wrong.

We route /projects/:id/badge specially, without a redirect, for speed.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>