Auto-locale: Completely rework for scope (:locale)
Change the approach to automatically determining the locale.
In particular, include almost everything in a scope of (:locale);
this makes the _url and _path helpers easier to work with.
This completely reworks config/routes.rb, while depending on
locale redirection being implemented in ApplicationController.
This sticks with the general recommendations for Rails i18n.

At this point there are still many failing tests.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>