Automatically determine locale (#1078)
* Automatically determine locale

We want users to automatically see their preferred locale.
This changes how URLs are interpreted; "no locale in URL"
is now interpreted as "redirect based on browser preference"
and not as "use English".

Spanish isn't ready, so do not automatically switch to its locale.

It's not clear what status code to provide on a locale-based redirect.
However, we must avoid 301 (Moved Permanently), because it is certainly
not a permanent move.  For the moment we'll use 300 (Multiple Choices),
because that code indicates there's a redirect based on agent choices
(which is certainly true).  We have to set the status to *something*,
because the Rails default is 301 and that is absolutely wrong.

We route /projects/:id/badge specially, without a redirect, for speed.

Include almost everything in a scope of (:locale);
this makes the _url and _path helpers easier to work with.
This completely reworks config/routes.rb, while depending on
locale redirection being implemented in ApplicationController.
This sticks with the general recommendations for Rails i18n.

* Auto-locale: Fix many tests

This commit has many changes, but almost all of them are
tiny tweaks to tests because the interface has changed slightly
(users must specify their locale, else be redirected).

* Don't redirect locale if JSON format requested

Handle a special case: If the requested format is JSON, don't bother
redirecting to the "best" locale, because JSON is the same in any locale.
This should make ease transition
(because the previous behavior for JSON files is completely unchanged).

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>