Add link alternate and publisher metadata
This adds link rel="alternate" metadata to make it much
easier to find the alternate locales for various pages.
It includes an "x-default" locale, which eases
automatically finding the "no locale" URL.

It also adds link rel="publisher" so that it's clear
who the publisher is.

This should help search engines find (and understand) the site.
That's especially important now that the site is automatically
redirecting based on locale.

For more information, see:
https://support.google.com/webmasters/answer/189077?hl=en
https://searchengineland.com/
the-ultimate-guide-to-multilingual-and-multiregional-seo-157838

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>