Add tests for locale links, update static_pages test
Add some tests for the locale-related links.

There's a weird test environment artifact I haven't been
able to track down.  The view response sometimes has an original url of
http://127.0.0.1:31337 and other times it's http://www.example.com.
Values such as "request.host" are consistently the second value.
This doesn't happen when we only test this file, but instead happens
when there's a full "rails test" - which means some other test
causes this.  It seems to be an artifact of the test environment, and
not actually a bug in the deployed code, so the test here will be
flexible to handle the variations that occur in the test environment.
This odd situation is noted in comments within the tests.

These regexes are defined using the %r{...}x format.
They are unusually long, and the "x" format makes them easier-to-read.

This also converts the StaticPagesControllerTest
from the deprecated "ActionController::TestCase" to
"ActionDispatch::IntegrationTest".  That didn't fix the
artifacts, but it seemed like a good thing to do anyway.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>