Select and use a canonical URL form
Select/use/document a canonical URL form.  We'll follow the
Rails path/url helper convention for the URL path:
at the root (without a locale) have a trailing "/", otherwise do not.
Also, we always generate locales as the first part of the URL path.
It's better for search engine optimization to choose
a canonical URL form.  Even in cases where it doesn't matter,
always having a canonical form makes it easier to detect problems.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>