Add link alternate and publisher metadata (#1084)
* Add link alternate and publisher metadata

This adds link rel="alternate" metadata to make it much
easier to find the alternate locales for various pages.
It includes an "x-default" locale, which eases
automatically finding the "no locale" URL.

It also adds link rel="publisher" so that it's clear
who the publisher is.

This should help search engines find (and understand) the site.
That's especially important now that the site is automatically
redirecting based on locale.

For more information, see:
https://support.google.com/webmasters/answer/189077?hl=en
https://searchengineland.com/
the-ultimate-guide-to-multilingual-and-multiregional-seo-157838

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>

* Add tests for locale links, update static_pages test

Add some tests for the locale-related links.

There's a weird test environment artifact I haven't been
able to track down.  The view response sometimes has an original url of
http://127.0.0.1:31337 and other times it's http://www.example.com.
Values such as "request.host" are consistently the second value.
This doesn't happen when we only test this file, but instead happens
when there's a full "rails test" - which means some other test
causes this.  It seems to be an artifact of the test environment, and
not actually a bug in the deployed code, so the test here will be
flexible to handle the variations that occur in the test environment.
This odd situation is noted in comments within the tests.

These regexes are defined using the %r{...}x format.
They are unusually long, and the "x" format makes them easier-to-read.

This also converts the StaticPagesControllerTest
from the deprecated "ActionController::TestCase" to
"ActionDispatch::IntegrationTest".  That didn't fix the
artifacts, but it seemed like a good thing to do anyway.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>

* Select and use a canonical URL form

Select/use/document a canonical URL form.  We'll follow the
Rails path/url helper convention for the URL path:
at the root (without a locale) have a trailing "/", otherwise do not.
Also, we always generate locales as the first part of the URL path.
It's better for search engine optimization to choose
a canonical URL form.  Even in cases where it doesn't matter,
always having a canonical form makes it easier to detect problems.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>

* Redirect requests of a non-root URL with trailing '/'

This forces a canonical representation of URLs, by
redirecting a non-root URL with a trailing slash to the
same URL without a trailing slash.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>